Use a git submodule for the ERFA source code.
Now updating the ERFA source in static compilation is very simple.

This commit also modernises the call to bindgen, and pulls out all the
constants within ERFA (previously not all constants were found). Newly
included constants include ellipsoid types, represented as ints. To help
ensure that the type of these constants is what is needed to interface
with applicaable functions, a new cargo test is included.